requirements-dev.txt
-r requirements-testing.txt

# For compiling the stylesheets and JavaScript
django-pipeline
django-pipeline-compass-rubygem

# Only for debugging, but very useful and easiest to just have it here
django-debug-toolbar

# For akomantoso import
python-dateutil<2
lxml

# For migrations
South
popit-django==0.0.3
